+++ client/app/layout.js
@@ -35,7 +35,7 @@ export default function RootLayout({ children }) {
             </div>
             <div className="logo">
               <a href="#">
-                <img src="./images/icon/lenstudio-logo.svg" alt="logo" />
+                <img src="/images/icon/lenstudio-logo.svg" alt="logo" />
 
               </a>
             </div>
@@ -99,7 +99,7 @@ export default function RootLayout({ children }) {
                       <a href="#">
                         鏡頭
                         <span className="icon">
-                          <img src="./images/icon/arrow-down.svg" alt="Icon" />
+                          <img src="/images/icon/arrow-down.svg" alt="Icon" />
                         </span>
                       </a>
                       <ul className="lens-brands">
