--- client/app/shopping-cart/_components/shopping-cart-step1/shopping-cart-step1.scss
@@ -1,120 +1,0 @@
-  .j-lessonImg img{
-    width: 600px;
-  }
-  .j-rentImg img {
-    width: 185px;
-    height: 199px;
-  }
-
-  .j-payStep {
-    padding: 60px;
-    border-radius: 20px;
-    background-color: #f6f6f6;
-  }
-
-  .j-pCount {
-    width: 323px;
-  }
-
-  .j-amount-text {
-    width: 40px;
-  }
-
-  .j-payStep .j-pTitle {
-    font-size: 24px;
-  }
-
-  .j-ifCouponUse {
-    font-size: 16px;
-  }
-
-  .j-ifCouponUse input {
-    display: none;
-  }
-
-  .j-modal-title {
-    font-family: Inter;
-    font-size: 35px;
-    font-weight: 600;
-  }
-
-  .j-modal-body {
-    width: 99%;
-    border-radius: 20px;
-    border: 1px solid rgba(20, 49, 70, 0.24);
-    background: rgba(201, 201, 201, 0.00);
-  }
-
-  .j-modelButton button {
-    width: 350px;
-    height: 50px;
-    align-items: center;
-    border-radius: 50px;
-    background: #003150;
-    color: #FFF;
-  }
-
-  .j-modelButton button {
-    border: none;
-    background: #003150;
-
-    &:hover,
-    &:active {
-      background-color: #003150;
-      color: #FFF;
-    }
-  }
-
-  .btn-close {
-    left: 590px
-  }
-
-  .footer {
-    display: none;
-  }
-
-  @media(max-width:390px) {
-    .container {
-      display: flex;
-      flex-direction: column;
-    }
-
-    .container .j-shoppingCartTitle {
-      font-size: 65px;
-    }
-
-    .container .j-shoppingCartBox {
-      display: flex;
-      flex-direction: column;
-    }
-
-    .j-shoppingItemsMoblieBox {
-      width: 390px;
-    }
-
-    .j-cartItemBox {
-      width: 390px;
-      height: auto;
-    }
-
-    .j-lessonImg img {
-      width: 375px;
-    }
-
-    .footer {
-      display: block;
-    }
-
-    .copyright {
-      text-align: center;
-    }
-
-    .copyright .right-links {
-      display: none;
-    }
-
-    .copyright .right-links .copyRight {
-      display: block;
-    }
-
-  }
