--- client/app/shopping-cart/_components/shopping-cart-step2/shopping-cart-step2.scss
@@ -1,90 +1,0 @@
-  
- .j-cartItem .j-cameraImg {
-    width: 109px;
-    height: 89px;
-  }
-
-  .j-lessonImg img{
-    width: 600px;
-  }
-  
-  .j-rentImg img{
-    width: 185px;
-    height: 199px;
-  }
-
-  .j-rentCameraBrand{
-    font-family: "Noto Sans HK";
-    font-size: 12px;
-    font-style: normal;
-    font-weight: 400;
-    line-height: normal;
-  }
-
-  .j-amount-text {
-    width: 40px;
-  }
-
-  .j-payStep {
-    padding: 22px;
-    gap: 35px;
-    border-radius: 20px;
-    background-color: #f6f6f6;
-  }
-
-  .j-payStep .j-pTitle {
-    font-size: 24px;
-  }
-  .j-payTitle{
-    color: #003150;
-    font-family: Inter;
-    font-size: 24px;
-    font-style: normal;
-    font-weight: 700;
-    line-height: normal;
-  }
-  .j-buyerInput input{
-    border-radius: 20px;
-  }
-
-  @media(max-width:390px) {
-    .container {
-      display: flex;
-      flex-direction: column;
-    }
-
-    .container .j-shoppingCartBox {
-      display: flex;
-      flex-direction: column;
-    }
-
-    .shoppingItemsMoblieBox {
-      width: 390px;
-    }
-
-    .j-cartItemBox {
-      width: 390px;
-      height: auto;
-    }
-
-    .j-lessonImg img {
-      width: 375px;
-    }
-    
-    .footer {
-      display: block;
-    }
-
-    .copyright{
-      text-align: center;
-    }
-    
-    .copyright .right-links {
-      display: none;
-    }
-
-    .copyright .right-links .copyRight {
-      display: block;
-    }
-    
-  }
