--- client/app/shopping-cart/_components/shopping-cart-step3/shopping-cart-step3.scss
@@ -1,134 +1,0 @@
-  .j-shoppingItemsBox{
-    height: 720px !important;
-  }
-  .accordion-body{
-    background-color: #f6f6f6;
-  }
-
-  .j-lessonImg img{
-    width: 610px;
-  }
-  
-  .j-rentImg img{
-    width: 185px;
-    height: 199px;
-  }
-
-  .j-rentCameraBrand{
-    font-family: "Noto Sans HK";
-    font-size: 12px;
-    font-style: normal;
-    font-weight: 400;
-    line-height: normal;
-  }
-
-  .j-content .itemDetail {
-    gap: 15px;
-  }
-
-  .j-detialTypeContent {
-    width: 200px;
-    font-size: 14px;
-    color: #8F8F8F;
-  }
-
-  .amount-text {
-    width: 40px;
-  }
-
-  .addressTitle{
-    color: #000;
-    text-align: center;
-    font-family: "Noto Sans";
-    font-size: 48px;
-    font-style: normal;
-    font-weight: 500;
-    line-height: normal;
-  }
-  .addressDetailContent{
-    font-family: "Noto Sans";
-    font-size: 11px;
-    font-style: normal;
-    font-weight: 500;
-    line-height: normal;
-  }
-  .j-payStep {
-    width: 550px;
-    padding: 22px;
-    gap: 10px;
-    border-radius: 20px;
-    background-color: #f6f6f6;
-  }
-
-  .payTitle{
-    color: #003150;
-    font-family: Inter;
-    font-size: 24px;
-    font-style: normal;
-    font-weight: 700;
-    line-height: normal;
-  }
-  
-  .j-creditCardInput input{
-    border-radius: 20px;
-  }
-  .j-useCredit input{
-    display: none;
-  }
-  .j-needCheckBox span{
-    color: red;
-  }
-  .j-mustWrite{
-    color: red;
-  }
-
-  @media(max-width:390px) {
-    .container {
-      display: flex;
-      flex-direction: column;
-    }
-
-    .container .j-shoppingCartTitle {
-      font-size: 65px;
-    }
-
-    .container .j-shoppingCartBox {
-      display: flex;
-      flex-direction: column;
-    }
-
-    .j-shoppingItemsMoblieBox {
-      width: 390px;
-    }
-
-    .j-cartItemBox {
-      width: 390px;
-      height: auto;
-    }
-
-    .j-lessonImg img {
-      width: 375px;
-    }
-    .j-pay{
-      width: 390px;
-    }
-    .j-payStep {
-      width: 390px;
-    }
-    .footer {
-      display: block;
-    }
-
-    .copyright{
-      text-align: center;
-    }
-    
-    .copyright .right-links {
-      display: none;
-    }
-
-    .copyright .right-links .copyRight {
-      display: block;
-    }
-    
-  }
