+++ client/app/shopping-cart/page.js
@@ -1,18 +1,8 @@
 'use client'
 
-import ShoppingCartEmptyPage from './_components/shopping-cart-empty/page';
-import ShoppingCartOnePage from './_components/shopping-cart-step1/page';
-import ShoppingCartTwoPage from './_components/shopping-cart-step2/page';
-import ShoppingCartThreePage from './_components/shopping-cart-step3/page';
-import "./shopping-cart.scss";
 export default function ShoppingCartPage(props) {
   return (
     <>
-      <div>ShoppingCart Page</div>
-      <ShoppingCartEmptyPage></ShoppingCartEmptyPage>
-      <ShoppingCartOnePage></ShoppingCartOnePage>
-      <ShoppingCartTwoPage></ShoppingCartTwoPage>
-      <ShoppingCartThreePage></ShoppingCartThreePage>
     </>
   )
 }
