+++ client/app/shopping-cart/shopping-cart-step1/shopping-cart-step1.scss
@@ -1,0 +1,291 @@
+body {
+    height: 100vh;
+  }
+
+  body>footer {
+    position: sticky;
+    top: 100vh;
+  }
+
+  .j-publicFont {
+    font-family: "Noto Sans HK";
+    font-style: normal;
+    font-weight: 400;
+    line-height: normal;
+  }
+
+  hr {
+    color: #aBaBaB;
+  }
+
+  .j-shoppingCartTitleBox .j-shoppingCartTitle {
+    font-family: Inter;
+    font-size: 64px;
+    font-style: normal;
+    font-weight: 500;
+    line-height: normal;
+  }
+
+  .j-shoppingCartTitleBox small {
+    font-family: Inter;
+    font-size: 35px;
+    font-style: normal;
+    font-weight: 500;
+    line-height: normal;
+  }
+
+  .j-shoppingItemsBox {
+    overflow: auto;
+    height: 529px;
+  }
+
+  .j-cartItemBox {
+    border-radius: 20px;
+    background-color: #fff;
+  }
+
+  .j-cartItem {
+    border-radius: 20px;
+    background-color: #fff;
+    gap: 10px;
+  }
+
+  .j-cartItem .j-cameraImg {
+    width: 109px;
+    height: 89px;
+  }
+
+  .j-detailcollapse {
+    padding: 0;
+    font-size: 14px;
+    border: none;
+    background-color: #f6f6f6;
+    text-align: left;
+  }
+
+  .j-accBtn:not(.collapsed) {
+    box-shadow: none;
+    background: #f6f6f6;
+  }
+
+  .j-accitem {
+    border: none;
+  }
+
+  .j-accBtn {
+    background-color: #f6f6f6;
+  }
+
+  .j-accBtn:active {
+    background-color: #f6f6f6;
+  }
+
+  .j-accBtn:focus {
+    background-color: #f6f6f6;
+  }
+
+  .j-accBtn:visited {
+    background-color: #f6f6f6;
+  }
+
+  .j-accBtn:link {
+    background-color: #f6f6f6;
+  }
+
+  .j-accBody {
+    background-color: #f6f6f6;
+  }
+  .j-lessonImg img{
+    width: 600px;
+  }
+  .j-rentImg img {
+    width: 185px;
+    height: 199px;
+  }
+
+  .j-rentCameraBrand {
+    font-family: "Noto Sans HK";
+    font-size: 12px;
+    font-style: normal;
+    font-weight: 400;
+    line-height: normal;
+  }
+
+  .j-rentDeadLine {
+    color: red;
+  }
+
+  .j-content {
+    height: 108px;
+    gap: 44px;
+  }
+
+  .j-content .j-itemDetail {
+    gap: 15px;
+  }
+
+  .j-content .j-itemDetail .j-brand {
+    color: #8F8F8F;
+    font-size: 12px;
+  }
+
+  .j-content .j-itemDetail .j-model {
+    color: #000;
+    font-size: 14px;
+  }
+
+  .j-content .j-itemDetail .j-detail {
+    color: #000;
+    font-size: 14px;
+  }
+
+  .j-detialTypeContent {
+    width: 200px;
+    font-size: 14px;
+    color: #8F8F8F;
+  }
+
+  .j-payStep {
+    padding: 60px;
+    border-radius: 20px;
+    background-color: #fff;
+  }
+
+  .j-pCount {
+    width: 323px;
+  }
+
+  .j-amount-text {
+    width: 40px;
+  }
+
+  .j-Checkout {
+    height: 50px;
+    border-radius: 50px;
+    background: #003150;
+  }
+
+  .j-Checkout .btn {
+    color: #FFF;
+    font-size: 14px;
+  }
+
+  .j-payStep .j-pTitle {
+    font-size: 24px;
+  }
+
+  .j-ifCouponUse {
+    font-size: 16px;
+  }
+
+  .j-ifCouponUse input {
+    display: none;
+  }
+  .footer-block{
+    display: none;
+  }
+  .copyright {
+    background-color: #f6f6f6;
+    padding-right: 5%;
+    padding-left: 5%;
+  }
+
+  .copyright a,
+  .copyright span {
+    color: #000
+  }
+
+  .j-modal-title {
+    font-family: Inter;
+    font-size: 35px;
+    font-weight: 600;
+  }
+
+  .j-modal-body {
+    width: 99%;
+    border-radius: 20px;
+    border: 1px solid rgba(20, 49, 70, 0.24);
+    background: rgba(201, 201, 201, 0.00);
+  }
+
+  .j-modelButton button {
+    width: 350px;
+    height: 50px;
+    align-items: center;
+    border-radius: 50px;
+    background: #003150;
+    color: #FFF;
+  }
+
+  .j-modelButton button {
+    border: none;
+    background: #003150;
+
+    &:hover,
+    &:active {
+      background-color: #003150;
+      color: #FFF;
+    }
+  }
+
+  .j-decrease:active {
+    border: none;
+  }
+
+  .j-increase:active {
+    border: none;
+  }
+
+  .btn-close {
+    left: 590px
+  }
+
+  .footer {
+    display: none;
+  }
+
+  @media(max-width:390px) {
+    .container {
+      display: flex;
+      flex-direction: column;
+    }
+
+    .container .j-shoppingCartTitle {
+      font-size: 65px;
+    }
+
+    .container .j-shoppingCartBox {
+      display: flex;
+      flex-direction: column;
+    }
+
+    .j-shoppingItemsMoblieBox {
+      width: 390px;
+    }
+
+    .j-cartItemBox {
+      width: 390px;
+      height: auto;
+    }
+
+    .j-lessonImg img {
+      width: 375px;
+    }
+
+    .footer {
+      display: block;
+    }
+
+    .copyright {
+      text-align: center;
+    }
+
+    .copyright .right-links {
+      display: none;
+    }
+
+    .copyright .right-links .copyRight {
+      display: block;
+    }
+
+  }
