--- client/app/shopping-cart/shopping-cart.scss
@@ -1,175 +1,0 @@
-  body {
-    height: 100vh;
-    background-color: #fff;
-  }
-
-  hr {
-    color: #aBaBaB;
-  }
-
-  .j-publicFont {
-    font-family: "Noto Sans HK";
-    font-style: normal;
-    font-weight: 400;
-    line-height: normal;
-  }
-
-  .j-shoppingCartTitleBox {
-    .j-shoppingCartTitle{
-      font-family: Inter;
-      font-size: 90px;
-      font-style: normal;
-      font-weight: 500;
-      line-height: normal;
-    }
-
-    .small {
-      font-family: Inter;
-      font-size: 18px;
-      font-style: normal;
-      font-weight: 500;
-      line-height: normal;
-    }
-  }
-
-  .j-shoppingItemsBox {
-    overflow: auto;
-    height: 529px;
-  }
-
-  .j-cartItemBox {
-    border-radius: 20px;
-    background-color: #f6f6f6;
-  }
-
-  .j-cartItem {
-    border-radius: 20px;
-    background-color: #f6f6f6;
-    gap: 10px;
-  }
-
-  .j-cartItem .j-cameraImg {
-    width: 109px;
-    height: 89px;
-  }
-
-  .j-detailcollapse {
-    padding: 0;
-    font-size: 14px;
-    border: none;
-    background-color: #f6f6f6;
-    text-align: left;
-  }
-
-  .accordion-item{
-    border: none;
-  }
-
-  .j-accitem {
-    border: none;
-  }
-
-  .j-accBtn:not(.collapsed) {
-    box-shadow: none;
-    background: #f6f6f6;
-  }
-
-  .j-accBtn {
-    background-color: #f6f6f6;
-  }
-
-  .j-accBtn:active {
-    background-color: #f6f6f6;
-  }
-
-  .j-accBtn:focus {
-    background-color: #f6f6f6;
-  }
-
-  .j-accBtn:visited {
-    background-color: #f6f6f6;
-  }
-
-  .j-accBtn:link {
-    background-color: #f6f6f6;
-  }
-
-  .j-accBody {
-    background-color: #f6f6f6;
-  }
-
-  .j-rentCameraBrand {
-    font-family: "Noto Sans HK";
-    font-size: 12px;
-    font-style: normal;
-    font-weight: 400;
-    line-height: normal;
-  }
-
-  .j-rentDeadLine {
-    color: red;
-  }
-
-  .j-content {
-    height: 108px;
-    gap: 44px;
-  }
-
-  .j-content .j-itemDetail {
-    gap: 15px;
-  }
-
-  .j-content .j-itemDetail .j-brand {
-    color: #8F8F8F;
-    font-size: 12px;
-  }
-
-  .j-content .j-itemDetail .j-model {
-    color: #000;
-    font-size: 14px;
-  }
-
-  .j-content .j-itemDetail .j-detail {
-    color: #000;
-    font-size: 14px;
-  }
-
-  .j-detialTypeContent {
-    width: 200px;
-    font-size: 14px;
-    color: #8F8F8F;
-  }
-
-  .j-decrease:active {
-    border: none;
-  }
-
-  .j-increase:active {
-    border: none;
-  }
-  
-  .j-Checkout {
-    height: 50px;
-    border-radius: 50px;
-    background: #003150;
-  }
-
-  .j-Checkout .btn {
-    color: #FFF;
-    font-size: 14px;
-  }
-  .footer-block{
-    display: none;
-  }
-
-  .copyright {
-    
-    background-color: #fff;
-    padding-right: 5%;
-    padding-left: 5%;
-  }
-
-  .copyright a,
-  .copyright span {
-    color: #000
-  }
